requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.4
python-dotenv==1.0.1
pandas==2.2.3
numpy==1.26.4
streamlit==1.48.0
plotly==5.24.0
requests==2.32.3
httpx==0.28.1
pytest==8.2.2
pytest-cov==4.1.0
black==25.1.0
isort==5.13.2
flake8==7.1.1
matplotlib==3.9.1
pyyaml==6.0.1
pre-commit==4.3.0
